[Refactor] Member Converter 변경
